moved dataclasses to common dir
