+++ src/analysis/methods/skeletonPrivateNairSimilarity.py
@@ -1,6 +1,6 @@
 import numpy as np
 import pandas as pd
-from src.creation.immuneRepertoire import ImmuneRepertoire
+from src.entities import ImmuneRepertoire
 from src.creation.algorithms.simpleBetaDistance import simpleBetaDistance
 from src.creation.distance.levenshtein import levenshteinDistance
 from src.analysis.methods.skeletonPublicNairSimilarity import skeletonPublicNairSimilarity
