+++ src/analysis/methods/skeletonPublicNairSimilarity.py
@@ -1,12 +1,11 @@
 import numpy as np
 import pandas as pd
-from src.models import Repertoire
 from src.creation.algorithms.simpleBetaDistance import simpleBetaDistance
 from src.creation.distance.levenshtein import levenshteinDistance
 import igraph as ig
 
 from src.creation.utils.pathManager import pathManager
-from src.creation.immuneRepertoire import ImmuneRepertoire
+from src.entities import ImmuneRepertoire
 from src.factories import ImmuneRepertoireFactory
 
 path = pathManager().testDataPath / "healthy_test_clonotypes_0.csv"
