+++ src/creation/algorithms/simpleBetaDistance.py
@@ -5,8 +5,7 @@ import logging
 import pandas as pd
 from src.creation.algorithms.common_methods import *
 from src.creation.algorithms.algorithm import *
-from src.models import Network
-from src.creation.immuneNetwork import ImmuneNetwork
+from src.entities import ImmuneNetwork
 
 @algorithm
 def simpleBetaDistance(repertoire, distance, threshold = 0.8, **kwargs):
