+++ src/creation/algorithms/simpleVectorBetaDistance.py
@@ -6,7 +6,7 @@ from src.creation.algorithms.algorithm import *
 import pandas as pd
 from scipy.spatial.distance import pdist
 from scipy.spatial.distance import squareform
-from src.creation.immuneNetwork import ImmuneNetwork
+from src.entities import ImmuneNetwork
 
 from src.models import Network
