+++ src/creation/algorithms/simpleVectorDistance.py
@@ -7,7 +7,7 @@ import pandas as pd
 from scipy.spatial.distance import pdist
 from scipy.spatial.distance import squareform
 from src.models import Repertoire
-from src.creation.immuneNetwork import ImmuneNetwork
+from src.entities import ImmuneNetwork
 
 @algorithm
 def simpleVectorDistance( repertoire, distance, threshold = None, **kwargs):
