+++ src/creation/distance/levenshtein.py
@@ -4,13 +4,8 @@ from src.creation.algorithms.common_methods import numerizeTCRSeq
 import numpy as np
 import math
 
-import pandas as pd
 import numpy as np
-from src.creation.algorithms.simpleDistance import simpleDistance
-from src.creation.algorithms.common_methods import split_tcr_column
-import uuid
 from src.creation.utils.pathManager import pathManager
-from src.factories import ImmuneRepertoireFactory
 
 path = pathManager().testDataPath / "healthy_test_clonotypes_0.csv"
 
@@ -58,7 +53,7 @@ class levenshteinDistance:
 #     ref = "aaa"
 #     print(a.tcr_dist(x,ref))
 
-if __name__ == "__main__":
-    df_net = simpleDistance(repertoire=ImmuneRepertoireFactory.fromCSVTest(path), distance=levenshteinDistance(group = True))
-    print(df_net)
+# if __name__ == "__main__":
+#     df_net = simpleDistance(repertoire=ImmuneRepertoireFactory.fromCSVTest(path), distance=levenshteinDistance(group = True))
+#     print(df_net)
