+++ src/entities/immuneNetwork.py
