+++ src/entities/repertoireStats.py
@@ -3,7 +3,7 @@ import pandas as pd
 from src.creation.distance.hamming import hammingDistance 
 from src.creation.distance.levenshtein import levenshteinDistance 
 from src.creation.algorithms.simpleBetaDistance import simpleBetaDistance
-from src.creation.immuneRepertoire import ImmuneRepertoire
+from src.entities import ImmuneRepertoire
 from src.analysis.methods.skeletonPublicNairSimilarity import skeletonPublicNairSimilarity
 from src.analysis.methods.skeletonPrivateNairSimilarity import skeletonPrivateNairSimilarity
 from src.creation.algorithms.common_methods import *
@@ -15,7 +15,7 @@ from src.factories import ImmuneRepertoireFactory
 
 path = pathManager().testDataPath / "healthy_test_clonotypes_0.csv"
 
-class repertoireAnalysis:
+class RepertoireStats:
     def __init__(self,repertoire):
         repertoire.clones['tcra_aa'] = repertoire.clones['cdr3s_aa'].apply(lambda x: split_tcr_column(x, subunit="TRA"))
         repertoire.clones['tcrb_aa'] = repertoire.clones['cdr3s_aa'].apply(lambda x: split_tcr_column(x, subunit="TRB"))
@@ -43,6 +43,6 @@ class repertoireAnalysis:
         return [self.num_of_tcra, self.num_of_tcrb, self.num_of_all_tcr, self.unique_tcra_distribution, self.unique_tcrb_distribution, self.unique_all_tcr_distribution, self.simpson_index_tcra, self.simpson_index_tcrb, self.simpson_index_all_tcr, self.shannon_index_tcra, self.shannon_index_tcrb, self.shannon_index_all_tcr]
 
 if __name__ == "__main__":
-    repertoireStats =  repertoireAnalysis(ImmuneRepertoireFactory.fromCSVTest(path))
+    repertoireStats =  RepertoireStats(ImmuneRepertoireFactory.fromCSVTest(path))
     print(repertoireStats)
